refactor: перенести компонент NewsSection in folder modules
