+++ src/modules/NewsSection/new-big-card/__content/new-big-card__content.css
