+++ src/modules/NewsSection/new-big-card/__title/new-big-card__title.css
