+++ src/modules/NewsSection/new-big-card/new-big-card.css
